Release 0.0.2

+++ src/products/products.service.ts
@@ -5,13 +5,20 @@ import { Product } from 'src/schemas/Product.schema';
 import mongoose, { Model } from 'mongoose';
 import { UpdateProductDto } from './dto/UpdateProduct.dto';
 import { Shop } from 'src/schemas/Shop.schema';
+import { R2Service } from 'src/r2/r2.service';
 
 @Injectable()
 export class ProductsService {
-    constructor(@InjectModel(Product.name) private productModel: Model<Product>) {}
+    constructor(
+        @InjectModel(Product.name) private productModel: Model<Product>,
+        private readonly r2Service: R2Service
+    ) {}
     
-    createProduct(createProductDto: CreateProductDto, shop: Shop) {
-        const data = Object.assign(createProductDto, { shop: shop._id });
+    async createProduct(createProductDto: CreateProductDto, shop: Shop, image: Express.Multer.File) {
+        let url = ''
+        if (image) url = await this.r2Service.uploadFile(image.buffer, image.mimetype);
+
+        const data = Object.assign(createProductDto, { shop: shop._id, image_url: url});
 
         const createdProduct = new this.productModel(data);
         return createdProduct.save();
@@ -28,16 +35,30 @@ export class ProductsService {
         return this.productModel.findOne({ _id: id, shop: shop._id });
     }
 
-    deleteProduct(id: string, shop: Shop) {
+    async deleteProduct(id: string, shop: Shop) {
         const isValid = mongoose.Types.ObjectId.isValid(id);
         if (!isValid) throw new HttpException('Invalid ID', 400);
+
+        const product = await this.productModel.findOne({ _id: id, shop: shop._id });
+        if (!product) throw new HttpException('Product not found', 404);
+        if (product.image_url !== '') this.r2Service.deleteFile(product.image_url.split('/').pop())
         
         return this.productModel.findOneAndDelete({ _id: id, shop: shop._id });
     }
 
-    updateProduct(id: string, updateProductDto: UpdateProductDto, shop: Shop) {
+    async updateProduct(id: string, updateProductDto: UpdateProductDto, shop: Shop, image: Express.Multer.File) {
         const isValid = mongoose.Types.ObjectId.isValid(id);
         if (!isValid) throw new HttpException('Invalid ID', 400);
+
+        const product = await this.productModel.findOne({ _id: id, shop: shop._id });
+        if (!product) throw new HttpException('Product not found', 404);
+
+        if (product.image_url !== '') await this.r2Service.deleteFile(product.image_url.split('/').pop());   
+
+        if (image) {
+            let url = await this.r2Service.uploadFile(image.buffer, image.mimetype);
+            return this.productModel.findOneAndUpdate({ _id: id, shop: shop._id }, { ...updateProductDto, image_url: url }, { new: true });
+        } 
         
         return this.productModel.findOneAndUpdate({ _id: id, shop: shop._id }, updateProductDto, { new: true });
     }
